Introduce incremental parsing
Do not parse and lookup all text before the UI can be used.
Instead, start parsing text from the end, moving backwards, emitting
a signal and returning to the event loop between each match.

The new parsing behavior has a couple of benefits:

- it keeps all matches, including overlapping matches, instead of
  only the longest matches (which are not always correct).  The UI
  is not fully updated for this, but keyboard navigation does work.

- in the case that the application has a scrollable view of text, and
  it has reported the entire text content (not only the text on screen),
  the old parsing behavior would handle all content before thawing
  the UI which could be very slow.  Because the new parser begins at
  the end of the text and works backwards, and the scrollable view is
  hopefully likely to be at the end (e.g. a chat window with log), this
  will allow using the UI much earlier in some cases.  This is the
  primary motivation for the change.

And one significant downside: overall, the amount of time taken to
process all the text is much longer. So, this commit reduces latency
at the cost of throughput.

Fixes #15